@@ -1,33 +1,32 @@
 {
   "name": "clark",
   "displayName": "CLARK: Cybersecurity Labs and Resource Knowledge-base",
-  "version": "3.8.0",
+  "version": "4.8.0",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test:unit": "jest --maxWorkers=2 --verbose --coverage auth/ admin/ cube/details/ uri-retriever",
+    "test:unit": "jest --maxWorkers=2 --verbose --coverage uri-retriever featuredObjects",
     "test:lighthouse": "jest -c ./lighthouse/jest.config.lighthouse.js",
     "test:lighthouse:debug": "node --inspect node_modules/.bin/jest --watch -c ./lighthouse/jest.config.lighthouse.js --runInBand",
     "lint": "ng lint",
     "storybook": "start-storybook"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.16",
-    "@angular/cdk": "^7.3.0",
-    "@angular/cli": "^7.3.0",
-    "@angular/common": "^7.2.16",
-    "@angular/compiler": "^7.2.16",
-    "@angular/compiler-cli": "^7.2.16",
-    "@angular/core": "^7.2.16",
-    "@angular/forms": "^7.2.16",
-    "@angular/http": "^7.2.16",
-    "@angular/platform-browser": "^7.2.16",
-    "@angular/platform-browser-dynamic": "^7.2.16",
-    "@angular/router": "^7.2.16",
+    "@angular/animations": "^9.0.6",
+    "@angular/cdk": "^7.3.7",
+    "@angular/cli": "^9.0.6",
+    "@angular/common": "^9.0.6",
+    "@angular/compiler": "^9.0.6",
+    "@angular/compiler-cli": "^9.0.6",
+    "@angular/core": "^9.0.6",
+    "@angular/forms": "^9.0.6",
+    "@angular/platform-browser": "^9.0.6",
+    "@angular/platform-browser-dynamic": "^9.0.6",
+    "@angular/router": "^9.0.6",
     "@cyber4all/clark-taxonomy": "^4.0.1",
-    "aws-sdk": "^2.607.0",
+    "aws-sdk": "^2.638.0",
     "chance": "^1.1.4",
     "chart.js": "^2.9.3",
     "core-js": "^2.6.11",
@@ -35,41 +34,43 @@
     "install": "^0.12.2",
     "md5": "^2.2.1",
     "ng-click-outside": "^4.0.0",
-    "ng2-charts": "^1.6.0",
-    "ng2-ckeditor": "^1.2.6",
+    "ng2-charts": "^2.3.0",
+    "ng2-ckeditor": "^1.2.7",
     "ngx-cookie": "4.0.2",
     "ngx-virtual-scroller": "^1.0.11",
     "raven-js": "^3.22.3",
     "request-promise": "^4.2.5",
     "rxjs": "^6.5.4",
     "rxjs-compat": "^6.5.4",
-    "tslib": "^1.9.0",
+    "rxjs-tslint": "^0.1.7",
+    "ts-jest": "^25.2.1",
+    "tslib": "^1.10.0",
     "uuid": "^3.4.0",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/language-service": "^7.2.16",
-    "@babel/core": "^7.8.3",
-    "@storybook/angular": "^5.3.8",
+    "@angular-devkit/build-angular": "^0.900.6",
+    "@angular/language-service": "^9.0.6",
+    "@babel/core": "^7.8.7",
+    "@storybook/angular": "^5.3.14",
     "@types/aws-sdk": "^2.7.0",
     "@types/file-saver": "^1.3.0",
-    "@types/jest": "^23.1.1",
-    "@types/node": "^10.17.13",
-    "@types/uuid": "^3.4.6",
+    "@types/jest": "^23.3.14",
+    "@types/node": "^12.12.30",
+    "@types/uuid": "^3.4.8",
     "babel-loader": "^8.0.6",
     "canvas": "^2.6.1",
     "chrome-launcher": "^0.10.7",
-    "codelyzer": "^4.3.0",
+    "codelyzer": "^5.1.2",
     "install-peers": "^1.0.2",
     "jasmine-spec-reporter": "~4.2.1",
     "jest": "^24.9.0",
     "jest-preset-angular": "^6.0.2",
     "lighthouse": "^5.6.0",
-    "protractor": "~5.4.0",
+    "protractor": "^5.4.3",
     "ts-node": "~7.0.0",
     "tslint": "~5.10.0",
-    "typescript": "3.2.4"
+    "typescript": "3.7.5"
   },
   "prettier": {
     "singleQuote": true
