@@ -12,6 +12,10 @@ import {
 import { StatusDescriptions } from 'environments/status-descriptions';
 import { AuthService } from 'app/core/auth.service';
 import { LearningObject } from '@entity';
+import { environment } from '@env/environment';
+import { HttpHeaders, HttpClient } from '@angular/common/http';
+import { take, catchError } from 'rxjs/operators';
+import { of } from 'rxjs/internal/observable/of';
 
 @Component({
   selector: 'clark-learning-object-list-item',
@@ -35,16 +39,22 @@ export class LearningObjectListItemComponent implements OnChanges {
 
   statusDescription: string;
 
+  showChangeAuthor: boolean;
+
   // flags
   meatballOpen = false;
 
+  hasParents: boolean = false;
+
+  private headers = new HttpHeaders();
   constructor(
     private auth: AuthService,
     private statuses: StatusDescriptions,
-    private cd: ChangeDetectorRef
+    private cd: ChangeDetectorRef,
+    private http: HttpClient,
   ) {}
 
-  ngOnChanges(changes: SimpleChanges) {
+  async ngOnChanges(changes: SimpleChanges) {
     if (changes.status) {
       this.statuses
         .getDescription(
@@ -56,6 +66,7 @@ export class LearningObjectListItemComponent implements OnChanges {
           this.cd.detectChanges();
         });
     }
+    await this.checkForParents();
   }
 
   /**
@@ -68,12 +79,35 @@ export class LearningObjectListItemComponent implements OnChanges {
     this.meatballOpen = value;
   }
 
+  toggleChangeAuthorModal(value: boolean){
+    this.showChangeAuthor = value;
+  }
   /**
    * Check the logged in user's email verification status
    * @return {boolean} true if loggedin user has verified their email, false otherwise
    */
   get verifiedEmail(): boolean {
     return this.auth.user.emailVerified;
+  }
+
+  async checkForParents() {
+    const parentUri = `${environment.apiURL}/users/${encodeURIComponent(
+      this.learningObject.author.username
+      )}/learning-objects/${encodeURIComponent(
+      this.learningObject.id
+    )}/parents`;
+
+    await this.http.get(
+      parentUri,
+      { headers: this.headers, withCredentials: true }
+      ).pipe(
+      take(1),
+      catchError(e => of(e))
+    ).subscribe(object => {
+      if (object && object.length) {
+        this.hasParents = true;
+      }
+    });
   }
 
 }
