@@ -16,6 +16,7 @@ export class CollectionDetailsComponent implements OnInit, OnDestroy {
   key = new Subject<string>();
   collection;
   pictureLocation: string;
+  showContribute = false;
 
   COPY = {
     VIEWALL: 'View All'
@@ -24,7 +25,7 @@ export class CollectionDetailsComponent implements OnInit, OnDestroy {
   constructor(
     private route: ActivatedRoute,
     private collectionService: CollectionService,
-    private titleService: Title
+    private titleService: Title,
   ) {}
 
   ngOnInit() {
@@ -47,6 +48,9 @@ export class CollectionDetailsComponent implements OnInit, OnDestroy {
     this.key.next(this.collection.abvName);
     if (this.collection.abvName !== 'intro_to_cyber' && this.collection.abvName !== 'secure_coding_community' && this.collection.abvName !== 'plan c') {
       this.pictureLocation = '../../../assets/images/collections/' + this.collection.abvName + '.png';
+    }
+    if (this.collection.abvName === 'plan c') {
+      this.showContribute = true;
     }
     this.titleService.setTitle(this.collection.name + ' | CLARK');
   }
