@@ -4,6 +4,7 @@ import { LearningObjectService } from '../../learning-object.service';
 import { Query, OrderBy, SortType } from '../../../interfaces/query';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
+import { FeaturedObjectsService } from 'app/core/featuredObjects.service';
 
 @Component({
   selector: 'cube-featured',
@@ -21,8 +22,9 @@ export class FeaturedComponent implements OnInit {
     status: [LearningObject.Status.RELEASED]
   };
   loading = false;
+  collectionName: string;
 
-  constructor(private learningObjectService: LearningObjectService) {
+  constructor(private learningObjectService: LearningObjectService, private featureService: FeaturedObjectsService) {
     this.learningObjects = this.learningObjects.fill(new LearningObject());
   }
 
@@ -31,12 +33,16 @@ export class FeaturedComponent implements OnInit {
       this.loading = true;
       this.collection.pipe(takeUntil(this.destroyed$)).subscribe({
         next: collection => {
+          this.collectionName = collection;
           this.query.collection = collection;
           this.fetchLearningObjects();
         }
       });
     } else {
-      this.fetchLearningObjects();
+      this.featureService.getFeaturedObjects();
+      this.featureService.featuredObjects.subscribe(objects => {
+        this.learningObjects = objects;
+      });
     }
   }
 
