@@ -146,6 +146,7 @@ export class LearningOutcome {
     this._verb = (taxonomy.taxons[this.bloom] as { verbs: string[] }).verbs[0];
     this._text = '';
     this._mappings = [];
+
     if (outcome) {
       this.copyOutcome(outcome);
     }
@@ -155,10 +156,10 @@ export class LearningOutcome {
    * Copies properties of outcome to this outcome if defined
    *
    * @private
-   * @param {Partial<LearningOutcome>} outcome
+   * @param outcome
    * @memberof LearningOutcome
    */
-  private copyOutcome(outcome: Partial<LearningOutcome>): void {
+  private copyOutcome(outcome: any): void {
     if (outcome.id) {
       this.id = outcome.id;
     }
